CORE/game ~(reorganizing headers)
